+++ MythicIOGrabber.lua
@@ -105,7 +105,7 @@ local function mainEvents(_, event, ...)
 		if(C_LFGList.HasActiveEntryInfo()) then
 			if(miog.EntryCreation and not miog.EntryCreation:IsVisible()) then
 				local activeEntryInfo = C_LFGList.GetActiveEntryInfo()
-				local activityInfo = miog.requestActivityInfo(activeEntryInfo.activityIDs[1])
+				local activityInfo = miog:GetActivityInfo(activeEntryInfo.activityIDs[1])
 
 				LFGListEntryCreation_ClearAutoCreateMode(LFGListFrame.EntryCreation);